requirements.txt
instagrapi
python-dotenv
schedule
